ION, TEST PPT CLASE

+++ main.py
@@ -1,4 +1,4 @@
-numero1 = int(input("Introduce un número:"))
-numero2 = int(input("Introduce otro numero al que quieres elevar el anterior: "))
+numero1 = float(input("Introduce un número:"))
+numero2 = float(input("Introduce otro numero al que quieres elevar el anterior: "))
 resultado = numero1**numero2
 print("El resultado es:", resultado)
